mat-sed/requirements.txt
absl-py==2.3.1
arrow==1.4.0
attrs==25.3.0
audioread==3.0.1
backports-zoneinfo==0.2.1
cffi==1.17.1
click==8.1.8
cmake==4.2.1
codecarbon==2.1.4
continual-inference==1.2.4
contourpy==1.1.1
cryptography==46.0.5
cycler==0.12.1
cython==3.2.4
dcase-util==0.2.20
decorator==5.2.1
desed==1.3.7
einops==0.8.1
exceptiongroup==1.3.1
fonttools==4.57.0
future==1.0.0
fuzzywuzzy==0.18.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.1.0
jams==0.3.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy-dataset==0.0.15
librosa==0.9.2
lightning-utilities==0.11.9
lit==18.1.8
llvmlite==0.41.1
markdown==3.7
matplotlib==3.6.3
mir-eval==0.8.2
mutagen==1.47.0
numba==0.58.1
numpy==1.23.5
oauthlib==3.3.1
packaging==26.0
pandas==1.1.3
pathlib==1.0.1
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
protobuf==5.29.6
psds-eval==0.5.3
psutil==7.2.2
ptflops==0.7.3
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydot-ng==2.0.0
pyloudnorm==0.1.1
pynvml==11.5.3
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
pyyaml==6.0.3
referencing==0.35.1
requests-oauthlib==2.0.0
resampy==0.4.3
rpds-py==0.20.1
rsa==4.9.1
scaper==1.6.5
scikit-learn==1.3.2
scipy==1.10.1
sebbs==1.0.2
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.10.3.post1
sox==1.4.0
soxbindings==1.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==0.4.5
tomli==2.4.0
torchmetrics==1.5.2
tqdm==4.65.0
typing-extensions==4.13.2
validators==0.34.0
websockets==13.1
werkzeug==3.0.6
wget==3.2
yt-dlp==2024.10.22
zipp==3.20.2